Fixed character decomposition when ascii-ids is used
When ascii-ids variable is set asciidoc is expected to normalize
all non-ascii characters when creating id from section name.
Originally it uses "NFKD" method which does not decompose all unicode characters.
This commit solves that problem adding a list of characters with
its decomposed form for the characters that requires it.

Signed-off-by: Radoslaw Koppel <r.koppel@k-el.com>